do not log errors in shutdown that have been seen already
All non-fatal errors pass through the error handler first and are
handled there according to the error_reporting(). The shutdown function
is called always and get_last_error() may return an error we already
logged or ignored. However we are really only interested in errors
that were fatal and couldn't be handled earlier.

There is one notice left in core that gets thrown before our
error_handler is attached. I'll fix it in core.