handle race condition
When an event happens, it is

1) written to the file system
2) sent to the sentry server
3) deleted

It can happen that the indexer is triggered while 2) is in progress. The
indexer will see the event and will send it again (Sentry ignores
duplicate events). Meanwhile 2) completed and 3) was executed. When the
indexer is done sending, deletion will fail. Triggering a notice. We
suppress this notice now.