overwrite HTTP environment for JavaScript logs
All javascript errors are logged through lib/exe/ajax.php initializing
the HTTP context in there will make the errors seem to have originated
within the AJAX call, but that is not necessarily true.

This patch overwrites the HTTP context with only the data that is known
to JavaScript.

This patch also logs the original stack trace to an extra field as I
don't trust our Stacktrace parser, yet.